libobs: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about non-exhaustive switch statements
* Fixes warnings about variables shadowing pre-existing variables
* Fixes warnings about unused assignments

+++ libobs/audio-monitoring/osx/coreaudio-output.c
@@ -21,7 +21,7 @@ struct audio_monitor {
 	struct circlebuf empty_buffers;
 	struct circlebuf new_data;
 	audio_resampler_t *resampler;
-	size_t buffer_size;
+	uint32_t buffer_size;
 	size_t wait_size;
 	uint32_t channels;
